chore: Move ModelOutput to it's own directory. This will help with planned refactor for ModelOutput.
